型定義と型定義のファイル分割
