+++ src/components/Header.tsx
@@ -1,0 +1,2 @@
+import React from "react";
+import Link from "next/link";
