+++ src/pages/todos/[id].tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Link from "next/link";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 import {
