+++ src/types/Comment.tsx
@@ -1,0 +1,6 @@
+export type Comment = {
+  id?: string;
+  name: string;
+  content: string;
+  createdAt: Date;
+};
