+++ src/types/create.ts
@@ -1,0 +1,10 @@
+// todoの作成
+export type CreateFormProps = {
+    todoData: {
+      title: string;
+      content: string;
+      status: string;
+    };
+    handleInputChange: (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement >) => void;
+    addTodo: (e: React.FormEvent<HTMLFormElement>) => Promise<void>;
+};
